requirements.txt
boto3
Pillow
numpy
opencv-python
scikit-learn
